requirements.txt
Django==1.11.13
django-debug-panel==0.8.3
django-debug-toolbar==1.9.1
django-js-asset==1.1.0
django-mptt==0.9.0
Faker==0.8.15
ipaddress==1.0.22
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.7.3
pytz==2018.4
six==1.11.0
sqlparse==0.2.4
text-unidecode==1.2
